updates #2021 (#6)

+++ setup.sh
@@ -2,12 +2,12 @@
 
 readlink_bin="${READLINK_PATH:-readlink}"
 if ! "${readlink_bin}" -f test &> /dev/null; then
-  __DIR__="$(dirname "$(python -c "import os,sys; print(os.path.realpath(os.path.expanduser(sys.argv[1])))" "${0}")")"
+  __DIR__="$(dirname "$(python3 -c "import os,sys; print(os.path.realpath(os.path.expanduser(sys.argv[1])))" "${0}")")"
 else
   __DIR__="$(dirname "$("${readlink_bin}" -f "${0}")")"
 fi
 
-source "${__DIR__}/shlibs/functions.shlib"
+source "${__DIR__}/shlibs/functions.sh"
 
 set -E
 trap 'throw_exception' ERR
@@ -15,7 +15,7 @@ trap 'throw_exception' ERR
 #
 # versions
 #
-bash_completion_version=2.9
+bash_completion_version=2.11
 minishift_version=1.16.1
 
 uname="$(uname)"
@@ -31,6 +31,8 @@ install_bash_powerline() {
 }
 
 install_bash_completion() {
+  brew install autoconf automake
+
   cd /tmp
   curl -#LO "https://github.com/scop/bash-completion/releases/download/${bash_completion_version}/bash-completion-${bash_completion_version}.tar.xz"
   tar xf "bash-completion-${bash_completion_version}.tar.xz"
@@ -40,6 +42,8 @@ install_bash_completion() {
   make > /dev/null
   sudo make -s install
   rm -rf "bash-completion-${bash_completion_version}"*
+
+  mkdir -p ~/.bash_completion.d/
 }
 
 install_oc() {
@@ -73,20 +77,26 @@ read -p 'git-gpg: ' git_gpg
 ###############################################################################
 # symlink dropbox / dotfiles
 ###############################################################################
-mkdir -p ~/.docker ~/".vim/pack/${USER}/opt"
-curl -L#O https://github.com/dracula/vim/archive/master.zip
-unzip master.zip
-mv vim-master ~/".vim/pack/${USER}/opt/dracula"
-rm master.zip
+mkdir -p ~/.docker
 
-for f in ~/Dropbox/dotfiles/.*; do
-  if [[ "${f: -1}" == "." ]]; then
-    continue
-  fi
-  ln -sf "${f}" ~
-done
-chmod 600 ~/.ssh/id_*
-chmod 644 ~/.ssh/*.pub
+if [[ ! -d ~/".vim/pack/${USER}/opt/dracula" ]]; then
+  mkdir -p ~/".vim/pack/${USER}/opt/dracula"
+  curl -L#O https://github.com/dracula/vim/archive/master.zip
+  unzip master.zip
+  mv vim-master ~/".vim/pack/${USER}/opt/dracula"
+  rm master.zip
+fi
+
+if [[ -d ~/Dropbox/dotfiles/ ]]; then
+  for f in ~/Dropbox/dotfiles/.*; do
+    if [[ "${f: -1}" == "." ]]; then
+      continue
+    fi
+    ln -sf "${f}" ~
+  done
+  chmod 600 ~/.ssh/id_*
+  chmod 644 ~/.ssh/*.pub
+fi
 
 ################################################################################
 # install various tools
@@ -97,11 +107,11 @@ if [[ ! -f ~/.bash-powerline.sh ]]; then
   install_bash_powerline
 fi
 
-if [[ ! -z "${DOTFILES_INSTALL_OC}" ]]; then
+if [[ -n "${DOTFILES_INSTALL_OC}" ]]; then
   install_oc
 fi
 
-if [[ ! -z "${DOTFILES_INSTALL_MINISHIFT}" ]]; then
+if [[ -n "${DOTFILES_INSTALL_MINISHIFT}" ]]; then
   install_minishift
 fi
 
@@ -114,7 +124,11 @@ mkdir -p \
   ~/.ssh/config.d/
 
 if [[ ! -f ~/.ssh/id_ed25519 ]]; then
-  ssh-keygen -o -a 100 -t ed25519 -f ~/.ssh/id_ed25519
+  ssh-keygen -o -a 100 -t ed25519 -f ~/.ssh/id_ed25519 -C "${git_email}"
+fi
+
+if [[ ! -f ~/.ssh/id_rsa ]]; then
+  ssh-keygen -m PEM -a 100 -t rsa -b 4096 -f ~/.ssh/id_rsa -C "${git_email}"
 fi
 
 if [[ ! -f ~/.ssh/config ]]; then
@@ -153,14 +167,28 @@ if [[ ! -d ~/.bash_completion.d ]]; then
   mkdir -p \
     ~/.bash_completion.d
 
-  ln -sf "${__DIR__}"/.bash_completion.d/ssh ~/.bash_completion.d/ssh
+  if [[ ! -f ~/.bash_completion.d/ssh ]]; then
+    ln -sf "${__DIR__}"/.bash_completion.d/ssh ~/.bash_completion.d/ssh
+  fi
 
-  if command -v oc > /dev/null; then
+  if [[ ! -f ~/.bash_completion.d/oc ]] && command -v oc > /dev/null; then
     oc completion bash > ~/.bash_completion.d/oc
   fi
 
-  if command -v minishift > /dev/null; then
+  if [[ ! -f ~/.bash_completion.d/minishift ]] && command -v minishift > /dev/null; then
     minishift completion bash > ~/.bash_completion.d/minishift
+  fi
+
+  if [[ ! -f ~/.bash_completion.d/hal ]] && command -v hal > /dev/null; then
+    hal --print-bash-completion > ~/.bash_completion.d/hal
+  fi
+
+  if [[ ! -f ~/.bash_completion.d/kubectl ]] && command -v kubectl > /dev/null; then
+    kubectl completion bash > ~/.bash_completion.d/kubectl
+  fi
+
+  if [[ ! -f ~/.bash_completion.d/minikube ]] && command -v minikube > /dev/null; then
+    minikube completion bash > ~/.bash_completion.d/minikube
   fi
 fi
 
@@ -172,15 +200,15 @@ install_showallfiles
 ################################################################################
 if [[ "${uname}" == "Darwin" ]]; then
   if [[ ! -d ~/Library/Application\ Support/TextMate/Bundles/Strip-Whitespace-On-Save.tmbundle ]]; then
-    git clone git@github.com:bomberstudios/Strip-Whitespace-On-Save.tmbundle.git ~/Library/Application\ Support/TextMate/Bundles/Strip-Whitespace-On-Save.tmbundle
+    git clone https://github.com/bomberstudios/Strip-Whitespace-On-Save.tmbundle.git ~/Library/Application\ Support/TextMate/Bundles/Strip-Whitespace-On-Save.tmbundle
   fi
 
   if [[ ! -d ~/Library/Application\ Support/TextMate/Bundles/Whitespace.tmbundle ]]; then
-    git clone git@github.com:mads-hartmann/Whitespace.tmbundle.git ~/Library/Application\ Support/TextMate/Bundles/Whitespace.tmbundle
+    git clone https://github.com/mads-hartmann/Whitespace.tmbundle.git ~/Library/Application\ Support/TextMate/Bundles/Whitespace.tmbundle
   fi
 
   if [[ ! -d ~/Library/Application\ Support/Textmate/Bundles/Ensure-New-Line-at-the-EOF.tmbundle ]]; then
-    git clone git@github.com:hajder/Ensure-New-Line-at-the-EOF.tmbundle.git ~/Library/Application\ Support/Textmate/Bundles/Ensure-New-Line-at-the-EOF.tmbundle
+    git clone https://github.com/hajder/Ensure-New-Line-at-the-EOF.tmbundle.git ~/Library/Application\ Support/Textmate/Bundles/Ensure-New-Line-at-the-EOF.tmbundle
   fi
 fi
 
@@ -222,10 +250,15 @@ fi
 # brew (macos)
 ################################################################################
 if [[ "${uname}" == "Darwin" ]]; then
-  brew install jq
+  brew install git jq tfenv wget gettext make gh python3 nvm yarn --without-node
+
+  ln -sf /usr/local/opt/gettext/bin/envsubst /usr/local/bin/envsubst
+  ln -sf /usr/local/opt/make/libexec/gnubin/make /usr/local/bin/make
+
+  defaults write -g ApplePressAndHoldEnabled -bool true
 fi
 
 ################################################################################
-# install global as to not be affected by local dirt
+# pip
 ################################################################################
-sudo -H pip3 install ansible awscli boto3
+pip3 install --user ansible awscli boto3
